# api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Attendly API",
        "description": "교회 출석 관리 시스템 API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/api/admin/organization/villages/{villageId}": {
            "put": {
                "tags": [
                    "관리자-조직구조"
                ],
                "summary": "마을 수정",
                "description": "마을 정보를 수정합니다",
                "operationId": "updateVillage",
                "parameters": [
                    {
                        "name": "villageId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VillageUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVillageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/organization/gbs-groups/{gbsGroupId}": {
            "put": {
                "tags": [
                    "관리자-조직구조"
                ],
                "summary": "GBS 그룹 수정",
                "description": "GBS 그룹 정보를 수정합니다",
                "operationId": "updateGbsGroup",
                "parameters": [
                    {
                        "name": "gbsGroupId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GbsGroupUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseGbsGroupResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/organization/departments/{departmentId}": {
            "get": {
                "tags": [
                    "관리자-조직구조"
                ],
                "summary": "부서 조회",
                "description": "특정 부서 정보를 조회합니다",
                "operationId": "getDepartment",
                "parameters": [
                    {
                        "name": "departmentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseDepartmentResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "관리자-조직구조"
                ],
                "summary": "부서 수정",
                "description": "부서 정보를 수정합니다",
                "operationId": "updateDepartment",
                "parameters": [
                    {
                        "name": "departmentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DepartmentUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseDepartmentResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "관리자-조직구조"
                ],
                "summary": "부서 삭제",
                "description": "부서를 삭제합니다",
                "operationId": "deleteDepartment",
                "parameters": [
                    {
                        "name": "departmentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/admin/organization/villages": {
            "post": {
                "tags": [
                    "관리자-조직구조"
                ],
                "summary": "마을 생성",
                "description": "새로운 마을을 생성합니다",
                "operationId": "createVillage",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VillageCreateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVillageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/organization/reorganization": {
            "post": {
                "tags": [
                    "관리자-조직구조"
                ],
                "summary": "GBS 6개월 주기 재편성 실행",
                "description": "GBS 그룹의 6개월 주기 재편성을 실행합니다",
                "operationId": "executeGbsReorganization",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GbsReorganizationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseGbsReorganizationResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/organization/gbs-groups": {
            "post": {
                "tags": [
                    "관리자-조직구조"
                ],
                "summary": "GBS 그룹 생성",
                "description": "새로운 GBS 그룹을 생성합니다",
                "operationId": "createGbsGroup",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GbsGroupCreateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseGbsGroupResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/organization/gbs-groups/{gbsGroupId}/members": {
            "post": {
                "tags": [
                    "관리자-조직구조"
                ],
                "summary": "GBS 그룹에 조원 배정",
                "description": "GBS 그룹에 조원을 배정합니다",
                "operationId": "assignMemberToGbs",
                "parameters": [
                    {
                        "name": "gbsGroupId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GbsMemberAssignRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/organization/gbs-groups/{gbsGroupId}/leaders": {
            "post": {
                "tags": [
                    "관리자-조직구조"
                ],
                "summary": "GBS 그룹에 리더 배정",
                "description": "GBS 그룹에 리더를 배정합니다",
                "operationId": "assignLeaderToGbs",
                "parameters": [
                    {
                        "name": "gbsGroupId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GbsLeaderAssignRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/organization/departments": {
            "get": {
                "tags": [
                    "관리자-조직구조"
                ],
                "summary": "부서 목록 조회",
                "description": "모든 부서를 조회합니다",
                "operationId": "getAllDepartments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePageResponseDepartmentResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "관리자-조직구조"
                ],
                "summary": "부서 생성",
                "description": "새로운 부서를 생성합니다",
                "operationId": "createDepartment",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DepartmentCreateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseDepartmentResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "VillageUpdateRequest": {
                "required": [
                    "name"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 50,
                        "minLength": 2,
                        "type": "string"
                    },
                    "departmentId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "villageLeaderId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "ApiResponseVillageResponse": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "$ref": "#/components/schemas/VillageResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "GbsGroupUpdateRequest": {
                "required": [
                    "name"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 50,
                        "minLength": 2,
                        "type": "string"
                    },
                    "villageId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "termStartDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "termEndDate": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "ApiResponseGbsGroupResponse": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "$ref": "#/components/schemas/GbsGroupResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "DepartmentUpdateRequest": {
                "required": [
                    "name"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 50,
                        "minLength": 2,
                        "type": "string"
                    }
                }
            },
            "ApiResponseDepartmentResponse": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "$ref": "#/components/schemas/DepartmentResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "ApiResponseVoid": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "VillageCreateRequest": {
                "required": [
                    "departmentId",
                    "name",
                    "villageLeaderId"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 50,
                        "minLength": 2,
                        "type": "string"
                    },
                    "departmentId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "villageLeaderId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "GbsReorganizationRequest": {
                "required": [
                    "departmentId",
                    "endDate",
                    "startDate"
                ],
                "type": "object",
                "properties": {
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "departmentId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "ApiResponseGbsReorganizationResponse": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "$ref": "#/components/schemas/GbsReorganizationResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "GbsGroupCreateRequest": {
                "required": [
                    "name",
                    "termEndDate",
                    "termStartDate",
                    "villageId"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 50,
                        "minLength": 2,
                        "type": "string"
                    },
                    "villageId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "termStartDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "termEndDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "leaderId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "GbsMemberAssignRequest": {
                "required": [
                    "memberId",
                    "startDate"
                ],
                "type": "object",
                "properties": {
                    "memberId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "GbsLeaderAssignRequest": {
                "required": [
                    "leaderId",
                    "startDate"
                ],
                "type": "object",
                "properties": {
                    "leaderId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "DepartmentCreateRequest": {
                "required": [
                    "name"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 50,
                        "minLength": 2,
                        "type": "string"
                    }
                }
            },
...
            "ApiResponsePageResponseDepartmentResponse": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PageResponseDepartmentResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}